requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
openai>=1.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langchain-community>=0.3.0
tiktoken>=0.5.0
tavily-python>=0.3.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
lancedb==0.13.0
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0